requirements.txt
os
re
subprocess
notion.client
notion.block
pyautogui
time